+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/rest_api/models/place_multiple_orders_batch_orders_parameter_inner.py
@@ -113,9 +113,9 @@ class PlaceMultipleOrdersBatchOrdersParameterInner(BaseModel):
         if value is None:
             return value
 
-        if value not in set(["GTC", "IOC", "FOK", "GTX", "GTD"]):
+        if value not in set(["GTC", "IOC", "FOK", "GTX", "GTD", "RPI"]):
             raise ValueError(
-                "must be one of enum values ('GTC', 'IOC', 'FOK', 'GTX', 'GTD')"
+                "must be one of enum values ('GTC', 'IOC', 'FOK', 'GTX', 'GTD', 'RPI')"
             )
         return value