+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/websocket_streams/websocket_streams.py
@@ -124,6 +124,9 @@ class DerivativesTradingUsdsFuturesWebSocketStreams(WebSocketStreamBase):
 
                 The Aggregate Trade Streams push market trade information that is aggregated for fills with same price and taking side every 100 milliseconds. Only market trades will be aggregated, which means the insurance fund trades and ADL trades won't be aggregated.
 
+
+        Retail Price Improvement(RPI) orders are aggregated and without special tags to be distinguished.
+
         Update Speed: 100ms
 
                 Args:
@@ -148,6 +151,8 @@ class DerivativesTradingUsdsFuturesWebSocketStreams(WebSocketStreamBase):
                 All Book Tickers Stream
 
                 Pushes any update to the best bid or ask's price or quantity in real-time for all symbols.
+
+        Retail Price Improvement(RPI) orders are not visible and excluded in the response message.
 
         Update Speed: 5s
 
@@ -335,6 +340,8 @@ class DerivativesTradingUsdsFuturesWebSocketStreams(WebSocketStreamBase):
 
                 Bids and asks, pushed every 250 milliseconds, 500 milliseconds, 100 milliseconds (if existing)
 
+        Retail Price Improvement(RPI) orders are not visible and excluded in the response message.
+
         Update Speed: 250ms, 500ms, 100ms
 
                 Args:
@@ -363,6 +370,8 @@ class DerivativesTradingUsdsFuturesWebSocketStreams(WebSocketStreamBase):
                 Individual Symbol Book Ticker Streams
 
                 Pushes any update to the best bid or ask's price or quantity in real-time for a specified symbol.
+
+        Retail Price Improvement(RPI) orders are not visible and excluded in the response message.
 
         Update Speed: Real-time
 
@@ -594,6 +603,8 @@ class DerivativesTradingUsdsFuturesWebSocketStreams(WebSocketStreamBase):
                 Partial Book Depth Streams
 
                 Top **<levels\>** bids and asks, Valid **<levels\>** are 5, 10, or 20.
+
+        Retail Price Improvement(RPI) orders are not visible and excluded in the response message.
 
         Update Speed: 250ms, 500ms or 100ms